네비게이션바 각 버튼별 onclick이벤트 함수 추가
